Refactor(jira): Refactor IssueHandler2 to IssueHandler.
