+++ alert-issuetracker/src/main/java/com/synopsys/integration/alert/issuetracker/IssueCreator.java
@@ -22,6 +22,7 @@
  */
 package com.synopsys.integration.alert.issuetracker;
 
+//TODO revisit this to make sure it is correct
 public interface IssueCreator {
     IssueContentModel createContent(OperationType operation, String messageId);
 }
