+++ alert-issuetracker/src/main/java/com/synopsys/integration/alert/issuetracker/IssueHandler.java
@@ -27,41 +27,25 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.synopsys.integration.alert.common.SetMap;
-import com.synopsys.integration.alert.common.enumeration.ItemOperation;
 import com.synopsys.integration.alert.common.exception.AlertException;
-import com.synopsys.integration.alert.common.message.model.ComponentItem;
-import com.synopsys.integration.alert.common.message.model.LinkableItem;
-import com.synopsys.integration.alert.common.message.model.MessageContentGroup;
-import com.synopsys.integration.alert.common.message.model.ProviderMessageContent;
 import com.synopsys.integration.alert.issuetracker.config.IssueConfig;
 import com.synopsys.integration.alert.issuetracker.exception.IssueMissingTransitionException;
-import com.synopsys.integration.alert.issuetracker.message.IssueTrackerMessageParser;
 import com.synopsys.integration.alert.issuetracker.message.IssueTrackerResponse;
 import com.synopsys.integration.exception.IntegrationException;
 
-/**
- * @param <T> A class that represents a single issue.
- */
 public abstract class IssueHandler<T> {
     private final Logger logger = LoggerFactory.getLogger(getClass());
 
-    private IssueTrackerMessageParser issueTrackerMessageParser;
-
-    public IssueHandler(IssueTrackerMessageParser issueTrackerMessageParser) {
-        this.issueTrackerMessageParser = issueTrackerMessageParser;
-    }
-
-    public final IssueTrackerResponse createOrUpdateIssues(IssueConfig issueConfig, MessageContentGroup content) throws IntegrationException {
+    public final IssueTrackerResponse createOrUpdateIssues(IssueConfig issueConfig, Collection<IssueContentModel> contentModels) throws IntegrationException {
         Set<String> issueKeys = new HashSet<>();
-        for (ProviderMessageContent messageContent : content.getSubContent()) {
-            Set<String> issueKeysForMessage = createOrUpdateIssuesPerComponent(messageContent, issueConfig);
+        for (IssueContentModel messageContent : contentModels) {
+            Set<String> issueKeysForMessage = createOrUpdateIssuesPerComponent(issueConfig, messageContent);
             issueKeys.addAll(issueKeysForMessage);
         }
 
@@ -69,56 +53,35 @@ public abstract class IssueHandler<T> {
         return new IssueTrackerResponse(statusMessage, issueKeys);
     }
 
-    protected Set<String> updateIssueByTopLevelAction(IssueConfig issueConfig, String providerName, LinkableItem topic, LinkableItem nullableSubTopic, ItemOperation action) throws IntegrationException {
-        if (ItemOperation.DELETE == action) {
-            List<T> existingIssues = retrieveExistingIssues(issueConfig.getProjectKey(), providerName, topic, nullableSubTopic, null, null);
-            logger.debug("Attempting to resolve issues in the project {} for Provider: {}, Provider Project: {}[{}].", issueConfig.getProjectKey(), providerName, topic.getValue(), nullableSubTopic);
-            Set<T> updatedIssues = updateExistingIssues(existingIssues, issueConfig, providerName, topic.getName(), action, Set.of());
-            return updatedIssues
-                       .stream()
-                       .map(this::getIssueKey)
-                       .collect(Collectors.toSet());
-        } else {
-            logger.debug("The top level action was not a DELETE action so it will be ignored");
-        }
-        return Set.of();
-    }
-
-    protected Set<String> createOrUpdateIssuesByComponentGroup(IssueConfig issueConfig, String providerName, LinkableItem topic, LinkableItem nullableSubTopic, SetMap<String, ComponentItem> groupedComponentItems)
+    protected Set<String> createOrUpdateIssuesPerComponent(IssueConfig issueConfig, IssueContentModel contentModel)
         throws IntegrationException {
         Set<String> issueKeys = new HashSet<>();
         String projectName = issueConfig.getProjectName();
+        IssueProperties issueProperties = contentModel.getIssueProperties();
 
         SetMap<String, String> missingTransitionToIssues = SetMap.createDefault();
-        for (Set<ComponentItem> componentItems : groupedComponentItems.values()) {
-            try {
-                ComponentItem arbitraryItem = componentItems
-                                                  .stream()
-                                                  .findAny()
-                                                  .orElseThrow(() -> new AlertException(String.format("No actionable component items were found. Provider: %s, Topic: %s, SubTopic: %s", providerName, topic, nullableSubTopic)));
-                ItemOperation operation = arbitraryItem.getOperation();
-                String trackingKey = createAdditionalTrackingKey(arbitraryItem);
+        try {
+            OperationType operation = contentModel.getOperation();
 
-                List<T> existingIssues = retrieveExistingIssues(issueConfig.getProjectKey(), providerName, topic, nullableSubTopic, arbitraryItem, trackingKey);
-                logIssueAction(operation, projectName, providerName, topic, nullableSubTopic, arbitraryItem);
-                if (!existingIssues.isEmpty()) {
-                    Set<T> updatedIssues = updateExistingIssues(existingIssues, issueConfig, providerName, arbitraryItem.getCategory(), arbitraryItem.getOperation(), componentItems);
-                    updatedIssues
-                        .stream()
-                        .map(this::getIssueKey)
-                        .forEach(issueKeys::add);
-                } else if (ItemOperation.ADD == operation || ItemOperation.UPDATE == operation) {
-                    IssueContentModel contentModel = issueTrackerMessageParser.createIssueContentModel(OperationType.CREATE, trackingKey, providerName, topic, nullableSubTopic, componentItems, arbitraryItem);
-                    T issueModel = createIssue(issueConfig, providerName, topic, nullableSubTopic, arbitraryItem, trackingKey, contentModel);
-                    String issueKey = getIssueKey(issueModel);
-                    issueKeys.add(issueKey);
-                } else {
-                    logger.warn("Expected to find an existing issue, but none existed.");
-                }
-            } catch (IssueMissingTransitionException e) {
-                missingTransitionToIssues.add(e.getTransition(), e.getIssueKey());
+            List<T> existingIssues = retrieveExistingIssues(issueConfig.getProjectKey(), issueProperties);
+            logIssueAction(operation, projectName, issueProperties);
+            if (!existingIssues.isEmpty()) {
+                Set<T> updatedIssues = updateExistingIssues(existingIssues, issueConfig, contentModel);
+                updatedIssues
+                    .stream()
+                    .map(this::getIssueKey)
+                    .forEach(issueKeys::add);
+            } else if (OperationType.CREATE == operation || OperationType.UPDATE == operation) {
+                T issueModel = createIssue(issueConfig, contentModel);
+                String issueKey = getIssueKey(issueModel);
+                issueKeys.add(issueKey);
+            } else {
+                logger.warn("Expected to find an existing issue, but none existed.");
             }
+        } catch (IssueMissingTransitionException e) {
+            missingTransitionToIssues.add(e.getTransition(), e.getIssueKey());
         }
+
         if (!missingTransitionToIssues.isEmpty()) {
             StringBuilder missingTransitions = new StringBuilder();
             for (Map.Entry<String, Set<String>> entry : missingTransitionToIssues.entrySet()) {
@@ -126,61 +89,43 @@ public abstract class IssueHandler<T> {
                 missingTransitions.append(String.format("Unable to find the transition: %s, for the issue(s): %s", entry.getKey(), issues));
             }
 
-            String errorMessage = String.format("For Provider: %s. Project: %s. %s.", providerName, projectName, missingTransitions.toString());
+            String errorMessage = String.format("For Provider: %s. Project: %s. %s.", issueProperties.getProvider(), projectName, missingTransitions.toString());
             throw new AlertException(errorMessage);
         }
         return issueKeys;
     }
 
-    protected abstract T createIssue(IssueConfig issueConfig, String providerName, LinkableItem topic, LinkableItem nullableSubTopic, ComponentItem arbitraryItem,
-        String trackingKey, IssueContentModel contentModel) throws IntegrationException;
+    protected abstract T createIssue(IssueConfig issueConfig, IssueContentModel contentModel) throws IntegrationException;
 
-    protected abstract List<T> retrieveExistingIssues(String projectSearchIdentifier, String provider, LinkableItem topic, LinkableItem nullableSubTopic, ComponentItem componentItem, String alertIssueUniqueId)
-        throws IntegrationException;
+    protected abstract List<T> retrieveExistingIssues(String projectSearchIdentifier, IssueProperties issueProperties) throws IntegrationException;
 
-    protected abstract boolean transitionIssue(T issueModel, IssueConfig issueConfig, ItemOperation operation) throws IntegrationException;
+    protected abstract boolean transitionIssue(T issueModel, IssueConfig issueConfig, OperationType operation) throws IntegrationException;
 
     protected abstract void addComment(String issueKey, String comment) throws IntegrationException;
-
-    protected abstract String createAdditionalTrackingKey(ComponentItem componentItem);
 
     protected abstract String getIssueKey(T issueModel);
 
     protected abstract String getIssueTrackerUrl();
 
-    protected Set<T> updateExistingIssues(List<T> issuesToUpdate, IssueConfig issueConfig, String providerName, String category, ItemOperation operation, Set<ComponentItem> componentItems)
+    protected Set<T> updateExistingIssues(List<T> issuesToUpdate, IssueConfig issueConfig, IssueContentModel issueContentModel)
         throws IntegrationException {
         Set<T> updatedIssues = new HashSet<>();
         for (T issue : issuesToUpdate) {
             String issueKey = getIssueKey(issue);
             if (issueConfig.getCommentOnIssues()) {
-                List<String> operationComments = issueTrackerMessageParser.createOperationComment(providerName, category, operation, componentItems);
+                Collection<String> operationComments = issueContentModel.getAdditionalComments();
                 for (String operationComment : operationComments) {
                     addComment(issueKey, operationComment);
                 }
                 updatedIssues.add(issue);
             }
 
-            boolean didUpdateIssue = transitionIssue(issue, issueConfig, operation);
+            boolean didUpdateIssue = transitionIssue(issue, issueConfig, issueContentModel.getOperation());
             if (didUpdateIssue) {
                 updatedIssues.add(issue);
             }
         }
         return updatedIssues;
-    }
-
-    private Set<String> createOrUpdateIssuesPerComponent(ProviderMessageContent messageContent, IssueConfig issueConfig) throws IntegrationException {
-        String providerName = messageContent.getProvider().getValue();
-        LinkableItem topic = messageContent.getTopic();
-        LinkableItem nullableSubTopic = messageContent.getSubTopic().orElse(null);
-
-        Set<String> issueKeys;
-        if (messageContent.isTopLevelActionOnly()) {
-            issueKeys = updateIssueByTopLevelAction(issueConfig, providerName, topic, nullableSubTopic, messageContent.getAction().orElse(ItemOperation.INFO));
-        } else {
-            issueKeys = createOrUpdateIssuesByComponentGroup(issueConfig, providerName, topic, nullableSubTopic, messageContent.groupRelatedComponentItems());
-        }
-        return issueKeys;
     }
 
     private String createStatusMessage(Collection<String> issueKeys) {
@@ -192,11 +137,11 @@ public abstract class IssueHandler<T> {
         return String.format("Successfully created issue at %s. Issue Keys: (%s)", getIssueTrackerUrl(), concatenatedKeys);
     }
 
-    private void logIssueAction(ItemOperation operation, String issueTrackerProjectName, String providerName, LinkableItem topic, LinkableItem nullableSubTopic, ComponentItem arbitraryItem) {
-        String issueTrackerProjectVersion = nullableSubTopic != null ? nullableSubTopic.getValue() : "unknown";
-        String arbitraryItemSubComponent = arbitraryItem.getSubComponent().map(LinkableItem::getValue).orElse("unknown");
+    private void logIssueAction(OperationType operation, String issueTrackerProjectName, IssueProperties issueProperties) {
+        String issueTrackerProjectVersion = issueProperties.getSubTopicValue() != null ? issueProperties.getSubTopicValue() : "unknown";
+        String arbitraryItemSubComponent = issueProperties.getSubComponentValue() != null ? issueProperties.getSubTopicValue() : "unknown";
         logger.debug("Attempting the {} action on the project {}. Provider: {}, Provider Project: {}[{}]. Category: {}, Component: {}, SubComponent: {}.",
-            operation.name(), issueTrackerProjectName, providerName, topic.getValue(), issueTrackerProjectVersion, arbitraryItem.getCategory(), arbitraryItem.getComponent().getName(), arbitraryItemSubComponent);
+            operation.name(), issueTrackerProjectName, issueProperties.getProvider(), issueProperties.getTopicValue(), issueTrackerProjectVersion, issueProperties.getCategory(), issueProperties.getComponentValue(),
+            arbitraryItemSubComponent);
     }
-
 }
